Migrated and fixed test for 'make_gif' function
